io: enhance istream/bstreams with pipe to forked child
 * prunes the child pid to avoid zombies
 * handles the errors so e.g. file-not-found is reported properly

+++ src/ver.c
@@ -33,7 +33,7 @@ static int ver_indexes(struct apk_database *db, int argc, char **argv)
 			continue;
 
 		printf("%.*s [%s]\n",
-                       repo->description.len,
+                       (int) repo->description.len,
 		       repo->description.ptr,
 		       db->repos[i].url);
 	}
